Use ffmpeg for image conversion; make image format configurable (#32)
* chore(converter): update executable to use ffmpeg instead of cwebp

* refactor: Rename WebPConverter to ImageConverter and update method names

* feat: Update image conversion logic to use ffmpeg instead of cwebp

This change was made to improve the flexibility of the image conversion process, allowing output formats such as AVIF.

* refactor: Rename webp.py and update import statements

* refactor(config.json): replace cwebp_args with ffmpeg_args

* refactor(config.json): update image_quality to 65

This is a compromise between accuracy and file size that is good enough for both AVIF and WebP. However, 65 for AVIF is both more accurate and bigger than 65 for WebP.

https://tsev.dev/posts/2023-11-10-should-avif-be-the-dominant-image-format/
https://www.industrialempathy.com/posts/avif-webp-quality-settings/

* feat(config.json): add setting image_format

This commit modifies the config.json file to include a new property "image_format" with the default value "avif". This change makes it possible to choose the format to convert to.

* refactor(file_path_factory): Update image format to use config setting

This commit updates the FilePathFactory module to dynamically set the image extension based on the "image_format" property in the config.json file. The fallback value is "avif". This change allows for flexibility in choosing the format to convert images to.

* refactor(bulk conversion): make more flexible according to image_format set in config

* feat(config.json): add ffmpeg_args to optimize image conversion

"-thread 0" enables multithreading.
"-still-picture 1" is specific to the AV1 codec. It indicates that the input should be treated as a still picture, enabling optimizations for still images.

* docs(config.md): update to reflect changes in config.json

* create config properties for accessing image format

* return early if no images

* add return type

* use config property

* update config

restore cwebp args and default quality

* refactor(image_conversion): configurable image quality

The code now checks if the destination image format is webp and uses the image_quality from the config and applies the maximum compression level (6). For other formats (AVIF), image_quality must be translated to CRF. libaom for AVIF already uses very strong compression (--cpu-used=1) by default.

Refs: https://github.com/Ajatt-Tools/PasteImagesAsWebP/pull/32#issuecomment-2168951551

* refactor(image_conversion): move '-still-picture' argument

This arg is used by libaom but not by libwebp. Therefore it doesn't need to be in the general ffmpeg args but fits better in the specific AVIF args.

* refactor(image_conversion): remove '-still-picture' argument for animated or video formats

The '-still-picture' argument is not adequate if the source is animated. This change ensures that the argument is not included when converting from animated images or video files.

* refactor(image_conversion.py): use cwebp instead of ffmpeg for webp

* fix(image_conversion.py): add ffmpeg arguments from config

* fix(config.json): remove unnecessary comma in the ffmpeg_args array

* feat(image_conversion.py): ffmpeg 'scale' now behaves like cwebp '-resize'

* fix(image_conversion.py): handle "'TempFile' object has no attribute 'lower'" error

* fix(image_conversion.py): ensure resize_args is defined before use

This commit fixes the error by defining resize_args at the beginning of the _convert_image method, ensuring it's available for use in both the if and else branches of the conditional logic.

* add config property

* remove variable

* format

* add setter

* update ajt common

* add property image quality

* ffmpeg is not bundled

* refactor

* cache value

* add function quality_percent_to_avif_crf

* refactor

fix stringify_args

* move resize args

* fix message

* expand ffmpeg args

* simplify

* rename var

* raise if ffmpeg is not in PATH

* refactor bulk-convert

* update readme

---------

Co-authored-by: Ren Tatsumoto <tatsu@autistici.org>